server/item: move the item NBT codec out of nbtconv
Crossbow reaches nbtconv.WriteItem and nbtconv.MapItem through go:linkname
rather than an import, because nbtconv imports item and the import cannot go
back the other way.

A linkname pull creates no dependency edge, so the compiler never looks for
nbtconv and the linker only resolves the symbols if something else in the
binary happens to import it. Anything importing server (a real server) gets
it via server/entity and server/block, which is why this has not come up.
A binary that imports server/item on its own does not:

	package main

	import (
		"github.com/df-mc/dragonfly/server/item"
		"github.com/df-mc/dragonfly/server/world"
	)

	func main() {
		var nbt world.NBTer = item.Crossbow{}
		_ = nbt.EncodeNBT()
	}

	github.com/df-mc/dragonfly/server/item.Crossbow.EncodeNBT: relocation
	target github.com/df-mc/dragonfly/server/internal/nbtconv.WriteItem
	not defined

The stack codec is moved into item, where the type it serialises lives and
where Crossbow can call it directly. nbtconv.WriteItem, Item and MapItem stay
as one-line wrappers, so none of the 22 call sites in entity, block and
item/creative change, and both linknames and the blank unsafe import are gone.

The six map readers the codec needs are duplicated privately in item rather
than imported: nbtconv depends on item, so the codec has to live here for the
dependency to point the right way, and these are the only helpers it uses.

mapValue keeps its name and field names. gob identifies the type by name, so
worlds written before this still decode.
